feat: context-aware channel rate limit with float RPM support
- Context-aware skip (other channels available) or 429 (all rate-limited)
- RateLimitRPM/DefaultRPM changed from int to float64, window auto-scales
- controller/relay.go: skip rate-limited channels via ExcludeChannel
- middleware/distributor.go: remove reject, defer to controller
- constant/context_key.go: add rate limit tracking keys
- relaykit/types/error.go: add ErrorCodeChannelRateLimited
- model/channel_cache.go: add GetChannelIDsForGroupModel

+++ setting/operation_setting/channel_rate_limit_setting.go
@@ -3,9 +3,9 @@ package operation_setting
 import "github.com/QuantumNous/new-api/setting/config"
 
 type ChannelRateLimitSetting struct {
-	Enabled    bool `json:"enabled"`
-	DefaultRPM int  `json:"default_rpm"`
-	DefaultTPM int  `json:"default_tpm"`
+	Enabled    bool    `json:"enabled"`
+	DefaultRPM float64 `json:"default_rpm"`
+	DefaultTPM int     `json:"default_tpm"`
 }
 
 var channelRateLimitSetting = ChannelRateLimitSetting{
